# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Basic Authentication with username and password",
        "title": "URL Insight API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8090",
    "basePath": "/api/v1",
    "paths": {
        "/crawler/results": {
            "get": {
                "security": [
                    {
                        "JWTAuth": []
                    },
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crawler"
                ],
                "summary": "Get recent crawl results",
                "responses": {
                    "200": {
                        "description": "array of recent crawl results",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/crawler.CrawlResult"
                            }
                        }
                    }
                }
            }
        },
        "/crawler/workers": {
            "patch": {
                "security": [
                    {
                        "JWTAuth": []
                    },
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crawler"
                ],
                "summary": "Adjust crawler workers",
                "parameters": [
                    {
                        "enum": [
                            "add",
                            "remove"
                        ],
                        "type": "string",
                        "description": "Action (add or remove)",
                        "name": "action",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of workers to add/remove",
                        "name": "count",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "adjusted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/urls": {
            "get": {
                "security": [
                    {
                        "JWTAuth": []
                    },
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "List URLs (paginated)",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "example": 10,
                        "description": "page_size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated URL list",
                        "schema": {
                            "$ref": "#/definitions/model.PaginatedResponse-model_URLDTO"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWTAuth": []
                    },
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Create URL row",
                "parameters": [
                    {
                        "description": "URL to crawl",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.URLCreateRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "{id}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/urls/{id}": {
            "get": {
                "security": [
                    {
                        "JWTAuth": []
                    },
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Get one URL row",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "URL ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.URLDTO"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWTAuth": []
                    },
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Update URL row",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "URL ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "fields",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateURLInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWTAuth": []
                    },
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Delete URL row",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "URL ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "deleted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/urls/{id}/results": {
            "get": {
                "security": [
                    {
                        "JWTAuth": []
                    },
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Latest analysis snapshot + links",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "URL ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.URLResultsDTO"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/urls/{id}/start": {
            "patch": {
                "security": [
                    {
                        "JWTAuth": []
                    },
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Start crawl",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "URL ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 5,
                        "description": "Priority (1-10, default 5)",
                        "name": "priority",
                        "in": "query"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "queued",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/urls/{id}/stop": {
            "patch": {
                "security": [
                    {
                        "JWTAuth": []
                    },
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Stop crawl",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "URL ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "stopped",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "crawler.CrawlResult": {
            "type": "object",
            "properties": {
                "duration": {
                    "description": "Duration in nanoseconds",
                    "type": "integer",
                    "format": "int64",
                    "example": 1500000000
                },
                "error": {},
                "linkCount": {
                    "type": "integer"
                },
                "links": {
                    "description": "Optional: include the actual links if needed",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Link"
                    }
                },
                "status": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "urlid": {
                    "type": "integer"
                }
            }
        },
...
        "model.AnalysisResult": {
            "type": "object",
            "properties": {
                "broken_link_count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "external_link_count": {
                    "type": "integer"
                },
                "h1_count": {
                    "type": "integer"
                },
                "h2_count": {
                    "type": "integer"
                },
                "h3_count": {
                    "type": "integer"
                },
                "h4_count": {
                    "type": "integer"
                },
                "h5_count": {
                    "type": "integer"
                },
                "h6_count": {
                    "type": "integer"
                },
                "has_login_form": {
                    "type": "boolean"
                },
                "html_version": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "internal_link_count": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url_id": {
                    "type": "integer"
                }
            }
        },
        "model.Link": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "href": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_external": {
                    "type": "boolean"
                },
                "status_code": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "url_id": {
                    "type": "integer"
                }
            }
        },
        "model.PaginatedResponse-model_URLDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.URLDTO"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/model.PaginationMetaDTO"
                }
            }
        },
        "model.PaginationMetaDTO": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "totalItems": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "model.URLCreateRequestDTO": {
            "type": "object",
            "required": [
                "original_url"
            ],
            "properties": {
                "original_url": {
                    "type": "string",
                    "example": "https://example.com"
                }
            }
        },
        "model.URLDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "original_url": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "queued",
                        "running",
                        "done",
                        "error"
                    ]
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.URLResultsDTO": {
            "type": "object",
            "properties": {
                "analysis_results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AnalysisResult"
                    }
                },
                "links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Link"
                    }
                },
                "url": {
                    "$ref": "#/definitions/model.URLDTO"
                }
            }
        },
        "model.UpdateURLInput": {
            "type": "object",
            "properties": {
                "original_url": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "queued",
                        "running",
                        "done",
                        "error"
                    ]
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "JWTAuth": {
            "description": "JWT Authentication token, prefixed with \"Bearer \" followed by the token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}